fix(parameter,pwm,docs): 优化电流参数校验与PWM逻辑
- 新增FACTORY_OUTCUR_MAX_MA宏定义合法电流上限为10000mA
- 完善工厂用户数据的参数校验，拦截超出范围的电流配置值
- 更新协议层校验逻辑，替换硬编码阈值为统一宏定义
- 简化PWM值计算代码，移除不必要的强制类型转换
- 新增配置加载后的PWM重载调用，使变更立即生效
- 清理系统缓存文件，更新.gitignore忽略无关文件与构建日志
- 新增单元测试覆盖新增的校验规则与PWM重载流程
- 新增数字电源项目实施方案、开发规范与Keil环境测试文档

+++ Core/Src/sys_pwm.c
@@ -90,7 +90,7 @@ extern u8  fa_test_EN;
          }
          else
          {
-             pwm_value = (u32)(((u64)persent * (u32)SET_OUTCUR * (u32)PWM_USEFUL_RANGE) / ((u32)HWMAX_OUTCUR * 100U));
+             pwm_value = ((u32)persent * (u32)SET_OUTCUR * (u32)PWM_USEFUL_RANGE) / ((u32)HWMAX_OUTCUR * 100U);
          }
       }
       else   //产测模式
